init ci refactor
