+++ tests/tnt/concurrent_topology_test.go
@@ -1,9 +1,8 @@
-package tnt_test
+package tnt
 
 import (
 	"context"
 	"fmt"
-	"log"
 	"math/rand"
 	"sync"
 	"testing"
@@ -36,6 +35,9 @@ func (c *concurrentTopologyProvider) Init(tc vshardrouter.TopologyController) er
 		defer close(c.closed)
 		//nolint:errcheck
 		defer tc.AddReplicasets(ctx, removed)
+		// Hack until issue will be resolved: https://github.com/KaymeKaydex/go-vshard-router/issues/65
+		// A little pause to let finish NewRouter() with no err
+		time.Sleep(2 * time.Second)
 
 		type actiont int
 
@@ -110,11 +112,7 @@ func TestConncurrentTopologyChange(t *testing.T) {
 	1) Addreplicaset + Removereplicaset by random in one goroutine
 	2) Call ReplicaCall, MapRw and etc. in another goroutines
 	*/
-
-	if !isCorrectRun() {
-		log.Printf("Incorrect run of tnt-test framework")
-		return
-	}
+	skipOnInvalidRun(t)
 
 	// Don't run this parallel with other tests, because this test is heavy and used to detect data races.
 	// Therefore this test may impact other ones.
@@ -179,5 +177,7 @@ func TestConncurrentTopologyChange(t *testing.T) {
 	wg.Wait()
 
 	// is router.Close method required?
-	tc.Close()
+	// tc.Close()
+	// TODO: we removed the above close, because sometimes tests stuck because
+	// rs.conn.CloseGraceful() (in RemoveReplicaset) stucks due to some unknown reason yet.
 }
