+++ tests/tnt/doc.go
@@ -1,0 +1,1 @@
+package tnt
